requirements.txt
selenium
Pillow
google-genai
dotenv